New permissions per command using the allow argument in the botcmd decorator
The allow argument takes a callable with the signature (bot, msg) where bot is a
SleekBot and msg is a message dictionay as defined in SleekXMPP. It should return
True or False to indicate if the user can execute the command.

Additionally the denymsg decorator can used to add a custom deny message to either
the command or the permission checking callable (the first has precedence).

Look at admin.py form an example.