Users class is now mapped to a dictionary. Commandbot uses the User class declared in XML. Enum for groups.
Users are now stored in a dict(key = jid, value = group). This makes more sense that a set, as jids can be in only one of
those at a time. Groups are in an Enum to avoid typing errors. Membership checking and group count can be checked via
check and count functions. Properties reflecting the groups names are provided as syntactic sugar and backwards
compatibility.

TODO: expand Users and virtual set to accept addition, deletions and modifications of users. Use Users as a base from a
database backend user class